Split sections into components
