+++ lib/solidus_page_objects.rb
@@ -4,6 +4,11 @@ require 'site_prism'
 
 require 'solidus_page_objects/site_prism/element_checker'
 
+# require components
+require 'solidus_page_objects/components/main_navigation'
+require 'solidus_page_objects/components/search'
+require 'solidus_page_objects/components/header'
+
 # require sections
 require 'solidus_page_objects/sections/main_navigation'
 require 'solidus_page_objects/sections/search'
