+++ lib/solidus_page_objects/components/header.rb
@@ -1,0 +1,10 @@
+module SolidusPageObjects
+  module Components
+    module Header
+      def self.included(base)
+        base.section :search, Sections::Search, '#search-bar'
+        base.section :main_navigation, Sections::MainNavigation, '#main-nav-bar'
+      end
+    end
+  end
+end
