+++ lib/solidus_page_objects/sections/header.rb
@@ -1,8 +1,7 @@
 module SolidusPageObjects
   module Sections
     class Header < SitePrism::Section
-      section :search, Search, '#search-bar'
-      section :main_navigation, MainNavigation, '#main-nav-bar'
+      include Components::Header
     end
   end
 end
