+++ lib/solidus_page_objects/sections/main_navigation.rb
@@ -1,8 +1,7 @@
 module SolidusPageObjects
   module Sections
     class MainNavigation < SitePrism::Section
-      element :home_page_link, 'li#home-link a'
-      element :cart_link,      'li#link-to-cart a'
+      include Components::MainNavigation
     end
   end
 end
